Make all page types work with dynamic routes
