+++ components/TxnSCClose.js
@@ -1,48 +1,14 @@
 import React, { Component } from 'react'
 import { Table, Descriptions } from 'antd'
-import ReactMapboxGl, { Marker } from 'react-mapbox-gl'
 import Client from '@helium/http'
 import animalHash from 'angry-purple-tiger'
 
-const Mapbox = ReactMapboxGl({
-  accessToken:
-    'pk.eyJ1IjoicGV0ZXJtYWluIiwiYSI6ImNqMHA5dm8xbTAwMGQycXMwa3NucGptenQifQ.iVCDWzb16acgOKWz65AckA',
-})
+import dynamic from 'next/dynamic'
 
-const styles = {
-  selectedMarker: {
-    width: 14,
-    height: 14,
-    borderRadius: '50%',
-    backgroundColor: '#1B8DFF',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    border: '4px solid #fff',
-  },
-  transmittingMarker: {
-    width: 14,
-    height: 14,
-    borderRadius: '50%',
-    backgroundColor: 'black',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    border: '4px solid #fff',
-  },
-  gatewayMarker: {
-    width: 14,
-    height: 14,
-    borderRadius: '50%',
-    backgroundColor: '#A984FF',
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    border: '3px solid #8B62EA',
-    boxShadow: '0px 2px 4px 0px rgba(0,0,0,0.5)',
-    cursor: 'pointer',
-  },
-}
+const SCCloseMapbox = dynamic(() => import('../components/SCCloseMapbox'), {
+  ssr: false,
+  loading: () => <div />,
+})
 
 class TxnSCClose extends Component {
   state = {
@@ -104,7 +70,7 @@ class TxnSCClose extends Component {
 
   async loadData() {
     const list = await this.client.hotspots.list()
-    const allSpots = await list.take(10000)
+    const allSpots = await list.take(20000)
     this.setState({ hotspots: allSpots })
   }
 
@@ -149,31 +115,7 @@ class TxnSCClose extends Component {
 
     return (
       <div>
-        <Mapbox
-          style={`mapbox://styles/petermain/cjyzlw0av4grj1ck97d8r0yrk`}
-          container="map"
-          center={[-95.712891, 37.09024]}
-          containerStyle={{
-            height: '600px',
-            width: '100%',
-          }}
-          zoom={[3]}
-          movingMethod="jumpTo"
-        >
-          {txn.stateChannel.summaries.forEach((s) => {
-            const hotspot = hotspots.find((e) => e.address === s.client)
-            if (hotspot && hotspot.lng && hotspot.lat) {
-              return (
-                <Marker
-                  key={hotspot.address}
-                  style={styles.gatewayMarker}
-                  anchor="center"
-                  coordinates={[hotspot.lng, hotspot.lat]}
-                />
-              )
-            }
-          })}
-        </Mapbox>
+        <SCCloseMapbox hotspots={hotspots} txn={txn} />
 
         <Descriptions bordered>
           <Descriptions.Item label="Block Height" span={3}>
