+++ next.config.js
@@ -1,2 +1,10 @@
 const withImages = require('next-images')
-module.exports = withImages()
+const withCSS = require('@zeit/next-css')
+
+module.exports = withCSS(
+  withImages({
+    css: {
+      importAsGlobal: true,
+    },
+  }),
+)
