+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@zeit/next-css": "^1.0.1",
     "angry-purple-tiger": "1.0.5",
     "antd": "^4.1.4",
     "base64url": "3.0.1",
